Make GetDlgItemBaseInt() return a DWORD_PTR

+++ src/StyleEdit.c
@@ -131,7 +131,7 @@ INT_PTR CALLBACK StyleEditProc(HWND hwnd, UINT iMsg, WPARAM wParam, LPARAM lPara
 			switch(HIWORD(wParam))
 			{
 			case EN_CHANGE:
-				dwStyle = GetDlgItemBaseInt(hwnd, IDC_EDIT1, 16);
+				dwStyle = (DWORD)GetDlgItemBaseInt(hwnd, IDC_EDIT1, 16);
 
 				hwndList = GetDlgItem(hwnd, IDC_LIST1);
 
@@ -153,7 +153,7 @@ INT_PTR CALLBACK StyleEditProc(HWND hwnd, UINT iMsg, WPARAM wParam, LPARAM lPara
 
 		case IDC_APPLY:
 
-			dwStyle = GetDlgItemBaseInt(hwnd, IDC_EDIT1, 16);
+			dwStyle = (DWORD)GetDlgItemBaseInt(hwnd, IDC_EDIT1, 16);
 
 			if(state->fExtended)
 				SetWindowLong(state->hwndTarget, GWL_EXSTYLE, dwStyle);
@@ -189,7 +189,7 @@ INT_PTR CALLBACK StyleEditProc(HWND hwnd, UINT iMsg, WPARAM wParam, LPARAM lPara
 				
 				hwndList = GetDlgItem(hwnd, IDC_LIST1);
 
-				dwStyle  = GetDlgItemBaseInt(hwnd, IDC_EDIT1, 16);
+				dwStyle  = (DWORD)GetDlgItemBaseInt(hwnd, IDC_EDIT1, 16);
 
 				caretidx = (int)SendMessage(hwndList, LB_GETCARETINDEX, 0, 0);
 				cursel   = (int)SendMessage(hwndList, LB_GETSEL, caretidx, 0);
